better function name, move function in to utils, docs
